Move remaining PttController items into AudioTabController
